Remove app landing page
